+++ ls_Glass/core/changelog.lua
@@ -6,18 +6,14 @@ local _G = getfenv(0)
 
 -- Mine
 E.CHANGELOG = [[
-- Added support for the pet battle log.
-- Fixed an issue where custom fonts wouldn't apply on load.
-- Added Korean translation. Translated by netaras@Curse.
-- Added Portuguese translation. Translated by Azeveco@Curse.
-- Added Traditional Chinese translation. Translated by RainbowUI@Curse.
-- Updated French translation. Translated by Braincell1980@Curse.
-- Updated German translation. Translated by Solence1@Curse.
-
-### Known Issues
-
-- There's a bug where the chat frame sometimes stops updating properly after scrolling down with
-  the mouse wheel. It's a tricky one to fix, I'm still investigating it. But to resume the update
-  process you need to scroll up until you see the "Jump to Present"/"Unread Messages" button and
-  then click it.
+- Added options to change the edit box's position and offset.
+- Improved compatibility with other addons that re-format chat messages. IME, the most compatible
+  one is BasicChatMods by funkehdude, it has features my addon currently lacks, like button hiding.
+  That said, you definitely should disable overlapping features in other addons.
+- Replaced "Jump to Present" and "Unread Messages" with icons because those text strings were
+  obnoxiously long in some locales.
+- Fixed an issue where the chat frame would sometimes stop updating after scrolling down with the
+  mouse wheel.
+- Added Spanish translation. Translated by cacahuete_uchi@Curse.
+- Updated Korean translation. Translated by netaras@Curse and unrealcrom96@Curse.
 ]]